+++ src/Wino.Core.Domain/BasicTypesJsonContext.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Text.Json.Serialization;
-using Wino.Core.Domain.Models.Updates;
+using Wino.Core.Domain.Models.WhatsNew;
 
 namespace Wino.Core.Domain;
 
@@ -9,6 +9,5 @@ namespace Wino.Core.Domain;
 [JsonSerializable(typeof(int))]
 [JsonSerializable(typeof(List<string>))]
 [JsonSerializable(typeof(bool))]
-[JsonSerializable(typeof(UpdateNotes))]
-[JsonSerializable(typeof(List<UpdateNoteSection>))]
+[JsonSerializable(typeof(WhatsNewRelease))]
 public partial class BasicTypesJsonContext : JsonSerializerContext;